src/XGBoost/requirements.txt
pandera==0.8.1
Flask==2.0.2
jsonschema==4.3.2
matplotlib==3.5.1
joblib==1.1.0
PyYAML==6.0
scikit-learn==1.0.2
xgboost==1.5.1